requirements.txt
Django==5.1.6
psycopg2-binary==2.9.10
redis==4.5.5
celery
djangorestframework==3.15.2
requests==2.32.3
psutil==7.0.0
pillow
drf-yasg>=1.20.0
streamlink
python-vlc
yt-dlp
gevent==24.11.1
django-cors-headers
djangorestframework-simplejwt
m3u8
rapidfuzz==3.12.1
# PyTorch dependencies (CPU only)
--extra-index-url https://download.pytorch.org/whl/cpu/
torch==2.6.0+cpu

# ML/NLP dependencies
sentence-transformers==3.4.1
uwsgi
channels
channels-redis
daphne
